refactor: split mixed-concern files for structural coherence
- Extract getPageInfo from dom-queries.ts into focused page-info.ts module
- Extract log formatting (formatLogEntry, shouldCaptureLog) from communication.ts into log-formatting.ts
- Extract captureScreenshot from communication.ts into screenshot-capture.ts
- communication.ts is now a pure facade with only re-exports
- Split popup recording.ts: extract UI rendering helpers into recording-ui-state.ts
- recording.ts now owns only wiring/lifecycle, delegating rendering
