+++ src/lib/page-info.ts
@@ -1,0 +1,68 @@
+/**
+ * Purpose: Extracts comprehensive page metadata (viewport, headings, forms, links) for page-level context.
+ * Why: Separates page-level metadata extraction from per-element DOM querying so each can evolve independently.
+ * Docs: docs/features/feature/query-dom/index.md
+ */
+
+import { DOM_QUERY_MAX_TEXT } from './constants.js'
+
+// Page info result
+interface PageInfoResult {
+  url: string
+  title: string
+  viewport: { width: number; height: number }
+  scroll: { x: number; y: number }
+  documentHeight: number
+  headings: string[]
+  links: number
+  images: number
+  interactiveElements: number
+  forms: FormInfo[]
+}
+
+// Form info
+interface FormInfo {
+  id?: string
+  action?: string
+  fields: string[]
+}
+
+/**
+ * Get comprehensive page info
+ */
+export async function getPageInfo(): Promise<PageInfoResult> {
+  const headings: string[] = []
+  const headingEls = document.querySelectorAll('h1,h2,h3,h4,h5,h6')
+  for (const h of headingEls) {
+    headings.push((h.textContent || '').slice(0, DOM_QUERY_MAX_TEXT))
+  }
+
+  const forms: FormInfo[] = []
+  const formEls = document.querySelectorAll('form')
+  for (const form of formEls) {
+    const fields: string[] = []
+    const inputs = form.querySelectorAll('input,select,textarea')
+    for (const input of inputs) {
+      const inputEl = input as HTMLInputElement | HTMLSelectElement | HTMLTextAreaElement
+      if (inputEl.name) fields.push(inputEl.name)
+    }
+    forms.push({
+      id: form.id || undefined,
+      action: form.action || undefined,
+      fields
+    })
+  }
+
+  return {
+    url: window.location.href,
+    title: document.title,
+    viewport: { width: window.innerWidth, height: window.innerHeight },
+    scroll: { x: window.scrollX, y: window.scrollY },
+    documentHeight: document.documentElement.scrollHeight,
+    headings,
+    links: document.querySelectorAll('a').length,
+    images: document.querySelectorAll('img').length,
+    interactiveElements: document.querySelectorAll('button,input,select,textarea,a[href]').length,
+    forms
+  }
+}
